requirements.txt
streamlit
requests
PyYAML
pywebview
pystray
Pillow
